+++ packages/file-bundle/config/file-image.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-use Psr\Container\ContainerInterface;
 use Rekalogika\File\Image\ImageResizer;
 use Rekalogika\File\Image\ImageTwigExtension;
+use Rekalogika\File\Image\ImageTwigRuntime;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
@@ -21,15 +21,22 @@ use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
 
-    $services->set(ImageResizer::class)
+    $services
+        ->set(ImageResizer::class)
         ->tag('rekalogika.file.derivation.filter');
 
-    $services->set(ImageTwigExtension::class)
-        ->tag('container.service_subscriber')
+    $services
+        ->set(ImageTwigExtension::class)
         ->tag('twig.extension', [
             'priority' => 1000000,
+        ]);
+
+    $services
+        ->set(ImageTwigRuntime::class)
+        ->args([
+            service(ImageResizer::class),
         ])
-        ->call('setContainer', [
-            service(ContainerInterface::class),
+        ->tag('twig.runtime', [
+            'priority' => 1000000,
         ]);
 };